Add support for the idle command
Adds the GMpdIdle flags to represent the various subsystems
of MPD, as well as the GMpdIdleResponse class to implement
decoding of the flags.

GMpdClient is updated to implement the idle methods. In
order to prevent timeouts while the client is idle (i.e.
when the client is only expecting an idle response),
the timeout on the underlying socket is set to 0
and the ready times of the GSource's are set to -1.